Updated compiler flags and fixed some errors
Updated compiler flags to the same used by other `dyret` packages. Some
warnings had to be disabled since included sources do not adhere to
them.

+++ src/dynamixel_wrapper.cpp
@@ -219,7 +219,7 @@ bool setServoSpeeds(std::vector<int> servoIds, std::vector<float> servoSpeeds){
     return false;
   }
 
-  for (int i = 0; i < servoIds.size(); i++) {
+  for (size_t i = 0; i < servoIds.size(); i++) {
     int convertedSpeed = int(round(1023.0 * servoSpeeds[i]));
 
     uint8_t param_speed[4];
@@ -248,7 +248,7 @@ bool setServoSpeeds(std::vector<int> servoIds, std::vector<float> servoSpeeds){
 }
 
 void setServoAngles(std::vector<float> anglesInRad) {
-  for (int i = 0; i < anglesInRad.size(); i++) {
+  for (size_t i = 0; i < anglesInRad.size(); i++) {
     auto dynAngle = (int) round(((normalizeRad(anglesInRad[i]) / (2 * M_PI)) * 4095.0) + 2048.0);
 
     uint8_t param_goal_position[4];
@@ -273,7 +273,7 @@ void setServoAngles(std::vector<float> anglesInRad) {
 bool setServoPIDs(std::vector<int> servoIds, std::vector<float> servoPIDs){
   int dxl_comm_result;
 
-  for (int i = 0; i < servoIds.size(); i++){
+  for (size_t i = 0; i < servoIds.size(); i++){
     dxl_comm_result = packetHandler->write2ByteTxOnly(portHandler, (uint8_t) servoIds[i], ADDR_MX2_P_GAIN, static_cast<uint16_t>(trunc(servoPIDs[i*3] * FACT_MX2_P_GAIN)));     // Write P
     if (dxl_comm_result != COMM_SUCCESS){ printCommResult(dxl_comm_result, "P_GAIN"); return false; }
     dxl_comm_result = packetHandler->write2ByteTxOnly(portHandler, (uint8_t) servoIds[i], ADDR_MX2_I_GAIN, static_cast<uint16_t>(trunc(servoPIDs[(i*3)+1] * FACT_MX2_I_GAIN))); // Write I
